+++ R/activate.R
@@ -24,17 +24,21 @@ activate <- function(.data, what, ..., select_var = NULL) UseMethod("activate")
 #' @export
 activate.tidync <- function(.data, what, ..., select_var = NULL) {
   if (missing(what)) return(.data)
+  vargrids <- tidyr::unnest(.data$grid) 
+  
   what_name <- deparse(substitute(what))
   #print(what_name)
   #if (what_name %in% var_names(.data)) what <- what_name
-  if (what_name %in% .data$grid$variable) {
+  if (what_name %in% vargrids$variable) {
     ## use the variable to find the grid
-    what <- .data$grid$grid[.data$grid$variable == what_name]
+    what <- vargrids$grid[vargrids$variable == what_name]
     select_var <- what_name
-  } else if (what %in% .data$grid$variable){
+  } else if (what %in% .data$variable$name){
     select_var <- what
-    what <- .data$grid$grid[.data$grid$variable == what]
-    
+    if (!is.null(select_var)) vargrids <- vargrids[vargrids$variable == select_var, , drop = FALSE]
+    #what <- paste(paste0("D", .data$axis$dimension[.data$axis$variable == select_var]), collapse = ",")
+    #.data$grid$grid[.data$grid$variable == what]
+    what <- vargrids$grid[1L]
   }
 
   if (is.numeric(what)) {
@@ -48,10 +52,11 @@ activate.tidync <- function(.data, what, ..., select_var = NULL) {
   }
   active(.data) <- what
 
-  active_variables <- .data[["grid"]] %>% dplyr::filter(.data$grid == what) %>% 
+  active_variables <- vargrids %>% dplyr::filter(.data$grid == what) %>% 
     dplyr::inner_join(.data[["variable"]], c("variable" = "name"))
+  
   if (!is.null(select_var)) {
-    active_variables <- inner_join(active_variables, tibble::tibble(variable = select_var))
+    active_variables <- inner_join(active_variables, tibble::tibble(variable = select_var), "variable")
   }
   active_dimensions <- as.integer(gsub("^D", "", unlist(strsplit(active(.data), ","))))
   .data$dimension$active <- rep(FALSE, nrow(.data$dimension))