+++ tests/testthat/test-files-sources.R
@@ -17,7 +17,7 @@ test_that("files and bad files are handled", {
 
 test_that("verbs have a character method", {
   f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
-  l3file <- system.file("extdata/oceandata", f, package= "tidync")
+  l3file <- system.file("extdata/oceandata", f, package= "tidync", mustWork = TRUE)
   hyper_tibble(l3file, lon = lon < -140, lat = lat > 85) %>% expect_s3_class("tbl_df")
   hyper_filter(l3file, lon = lon < -140, lat = lat > 85) %>% expect_s3_class("tidync")
   hyper_slice(l3file, lon = lon < -140, lat = lat > 85, select_var = "chlor_a") %>% expect_type("list")